pyproject.toml
[project]
name = "spatial-atlas"
version = "1.0.0"
description = "Spatial-aware research agent built on compute-grounded reasoning (CGR) for AgentX-AgentBeats Phase 2 Sprint 2. Deterministic spatial scene graphs, entropy-guided model routing, and self-healing ML pipelines for field work analysis and Kaggle competition solving."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "httpx>=0.28.1",
    "litellm>=1.0.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.38.0",
    "pypdf>=5.0.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
